requirements.txt
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.9.1
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-channels==0.7.0
django-froala-editor==4.6.2
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==11.0.3
wsaccel==0.6.7
zope.interface==8.0.1
